+++ CHANGELOG.md
@@ -6,6 +6,7 @@ All notable changes to this project will be documented in this file.
 - Issue 29: Completed sample-mfg/start-mfg.sh bash script.
 - Converted swagger documentation to OpenAPI 3
 - README updates
+- Added `FIDO_DEVICE_ONBOARD_REL_VER` environment variable FDO Owner Service container and script.
 
 ## [1.1.0] - 2023-05-24
 - Issue 25: Updated FDO to support Intel's `1.1.5` release.